\name{TraME.R}

\alias{TraME.R}

\docType{package}

\title{Transportation Methods for Econometrics}

\description{
TraME (Transportation Methods for Econometrics) is an R package for for solving problems of equilibrium computation and estimation in consumer demand and matching frameworks via the Mass Transportation Approach.
}

\details{
\tabular{ll}{
Package: \tab TraME.R\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2017-08-13\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab Yes\cr
Depends: \tab Rcpp, RcppArmadillo, digest\cr
Suggests: \tab gurobi, Rglpk\cr}
}

\section{How to cite this package}{

 \preformatted{@Manual{Galichon2017,
 author       = {Alfred Galichon and Keith O'Hara},
 title        = {{TraME}: Transportation Methods for Econometrics},
 year         = {2017},
 note         = {R package version 0.1.0.}}}
\emph{}
}

\section{License}{
TraME is licensed under the GNU General Public License (GPL) version 2, or (at your option) any later version.
}

\author{Alfred Galichon}
\author{Keith O'Hara}
